refactor(rainix-static): split main.rs into per-subcommand modules
main.rs becomes a thin dispatcher (arg parse + crate::fail); the two subcommands
move to their own modules — no_submodules.rs (static check) and soldeer_gate.rs
(the Soldeer content gate). Pure refactor, no behavior change: the 14 unit tests
move with their code and still pass; clippy/fmt/reuse clean.
